refactor: move SbomType enum to core
